+++ package.json
@@ -28,6 +28,7 @@
     "packages/openapi-generator",
     "packages/generator-common",
     "packages/temporal-de-serializers",
+    "packages/middlewares",
     "test-packages/integration-tests",
     "test-packages/e2e-tests",
     "test-packages/test-services-odata-common",
@@ -69,6 +70,7 @@
     "http-client": "yarn workspace @sap-cloud-sdk/http-client",
     "mail-client": "yarn workspace @sap-cloud-sdk/mail-client",
     "openapi": "yarn workspace @sap-cloud-sdk/openapi",
+    "middlewares": "yarn workspace @sap-cloud-sdk/middlewares",
     "openapi-generator": "yarn workspace @sap-cloud-sdk/openapi-generator",
     "odata-common": "yarn workspace @sap-cloud-sdk/odata-common",
     "odata-v2": "yarn workspace @sap-cloud-sdk/odata-v2",
