+++ packages/http-client/src/http-client.spec.ts
@@ -36,8 +36,7 @@ import type {
   Middleware,
   MiddlewareIn,
   HttpMiddlewareContext
-} from './middleware/middleware-type';
-import { timeout } from './middleware/timeout';
+} from './middleware-type';
 
 describe('generic http client', () => {
   const httpsDestination: Destination = {
@@ -572,33 +571,6 @@ sap-client:001`);
       });
       await expect(promise).resolves.not.toThrow();
     }, 60000);
-
-    it('uses a custom timeout if given', async () => {
-      const delayInResponse = 10;
-      nock('https://example.com', {})
-        .get('/with-delay')
-        .times(2)
-        .delay(delayInResponse)
-        .reply(200);
-
-      await expect(
-        executeHttpRequest(httpsDestination, {
-          method: 'get',
-          url: '/with-delay',
-          middleware: [timeout(delayInResponse * 0.5)]
-        })
-      ).rejects.toThrow(
-        'Request to https://example.com ran into timeout after 5ms.'
-      );
-
-      await expect(
-        executeHttpRequest(httpsDestination, {
-          method: 'get',
-          url: '/with-delay',
-          middleware: [timeout(delayInResponse * 10)]
-        })
-      ).resolves.not.toThrow();
-    });
 
     it('overwrites the default axios config with destination related request config', async () => {
       const destination: Destination = {
