+++ packages/http-client/src/internal.ts
@@ -2,4 +2,4 @@ export * from './csrf-token-header';
 export * from './http-client';
 export * from './http-client-types';
 export * from './http-request-config';
-export * from './middleware';
+export * from './middleware-type';
