--- packages/http-client/src/middleware/index.ts
@@ -1,2 +1,0 @@
-export * from './timeout';
-export * from './middleware-type';
