+++ packages/middlewares/jest.config.js
@@ -1,0 +1,6 @@
+// eslint-disable-next-line
+const commonConfig = require('../../test-resources/jest.common.config');
+module.exports = {
+  ...commonConfig,
+  displayName: 'http-client'
+};
